AG-6548 - Move Charts API Explorer editor selection logic to Editors.tsx.
